Clean perception common (#2164)
* clear compiling warning in hough_transfer.cc

fixed some warnings in camera_homography.cc

fixed some warnings in conditional_clustering_test.cc

clearn warnings in basic.h

clear warnings in conditional_clustering.h

clear warnings in convex_hull_2d_test.cc

fixed a warning in convex_hull_2d.h

clean warnings in downsampling_test.cc

clean warnings in downsampling.h

clean warnings in i_plane.h

clean warnings in common.h

clean warnings in io_util.cc

fix bug induced when cleaning

* small modify of a fix

* address xiangquans comment

* address xiangquans comment

+++ modules/perception/common/geometry/convex_hull_2d.h
@@ -217,9 +217,9 @@ bool ConvexHull2D<CLOUD_IN_TYPE, CLOUD_OUT_TYPE>::GetConvexHullMonotoneChain(
   }
   out_polygon->clear();
   out_polygon->resize(polygon_indices_.size());
-  float min_z = in_cloud_->at(0).z;
+  float min_z = static_cast<float>(in_cloud_->at(0).z);
   for (std::size_t id = 0; id < in_cloud_->size(); ++id) {
-    min_z = std::min<float>(in_cloud_->at(id).z, min_z);
+    min_z = std::min<float>(static_cast<float>(in_cloud_->at(id).z), min_z);
   }
   for (std::size_t i = 0; i < polygon_indices_.size(); ++i) {
     out_polygon->at(i).x = points_[polygon_indices_[i]](0);
